requirements.txt
streamlit
dwave-ocean-sdk
numpy
